refactor: split humanReporter
